requirements.txt
PyQt5
black
mysql-connector-python
cryptography